Add files via upload
This is Chrome Extension "Athletic Timer". It is different than the normal timer as it has two modes "Standard Mode" and "Athletic Mode". The Standard mode works as a normal timer which can track time with pause, resume, reset functionalities whereas the, Athletic Mode has functionality in which it can measure the workout time , the break time and the number of repetitions the athlete performs, suitable for HIIT Training.